requirements.txt
altgraph
attrs
av
bcrypt==4.0.1
bidict
blinker
certifi
charset-normalizer
click
dotenv
Flask
flask-cors
Flask-Login
Flask-SocketIO
getmac
greenlet
h11
idna
isodate
itsdangerous
Jinja2
lxml
MarkupSafe
netifaces
numpy
onvif_zeep
opencv-python
packaging
pandas
passlib
pendulum
platformdirs
psutil
psycopg2-binary
pyinstaller
pyinstaller-hooks-contrib
python-dateutil
python-dotenv
python-engineio
python-socketio
pytz
requests
requests-file
requests-toolbelt
setuptools
simple-websocket
six
SQLAlchemy
typing_extensions
tzdata
urllib3
Werkzeug
wsproto
zeep